rebuild of point grey camera
made some changes to the python analysis side of things. I removed the "python analysis" variable from the "do not set" list of variable in initialization. I had put it there because in the configure case for python analysis, the "off" function closes an open python session, which would always through an error if a python session hadn't been opened. Now, upon initialization, the default is to have python analysis turned off, but I have a "first call" case added to the "off" setting that will do nothing to prevent the throwing of errors